feat(capabilities): ComponentHost peer-addressing facade (issue 654)
Moves `WasmTrampoline` from `aether-substrate::actor::wasm::trampoline`
to `aether-capabilities::trampoline` — next to the
`ComponentHostCapability` that spawns it — and applies the
`#[bridge(instanced)]` pattern. After the move,
`WasmTrampoline::NAMESPACE` is the single, cap-owned declaration of
the `aether.component.trampoline` prefix, reachable on every target
via the `Actor` trait const.

Why move the actor out of substrate: the trampoline's only
consumer-by-name in the workspace is `ComponentHostCapability`. Its
native impl needs `wasmtime` / `Registry` / `Mailer` / `HubOutbound`
— exactly what `aether-capabilities`'s `native` feature already
pulls. The bridge stub keeps the type reachable from wasm-component
senders so `ctx.actor::<ComponentHostCapability>().loaded::<R>(name)`
resolves the trampoline-prefixed address via
`WasmTrampoline::NAMESPACE` without depending on substrate.

ComponentHostFfiExt / ComponentHostNativeExt (the peer-addressing
facade from the issue) now read the namespace from
`WasmTrampoline::NAMESPACE` directly. No separate
`TRAMPOLINE_NAMESPACE` const, no `trampoline_address` helper, no
const-block byte-equality voodoo, no cross-crate duplicate.

Other moving pieces:

- `aether-actor` + `aether-substrate` — new inherent
  `resolve_peer<Peer>(name)` on `FfiActorMailbox<R>` /
  `NativeActorMailbox<'a, R>`. Lets the ext trait construct a sibling
  mailbox on the same transport without re-threading the ctx; the
  native variant inherits the parent's binding lifetime.
- `aether-substrate` — `ComponentCtx::install_binding` and
  `NativeInitCtx::binding()` promoted from `pub(crate)` to `pub`.
  Both methods existed specifically for the trampoline's use (the
  `install_binding` doc comment even names `WasmTrampoline::init` as
  the caller); promoting them is just acknowledging the actual usage
  pattern now that the consumer lives in a sibling crate.
- `aether-substrate/src/mail/registry.rs:619` — the routing
  categorisation `starts_with("aether.component.trampoline:")` keeps
  its literal (substrate can't depend on capabilities, wrong dep
  direction) with a comment linking back to `WasmTrampoline::NAMESPACE`
  as the source of truth.
- Test migrations — camera, mesh-viewer, substrate-bundle, sokoban
  scenarios switch from a hard-coded `"aether.component.trampoline:NAME"`
  to a helper built on `aether_capabilities::WasmTrampoline::NAMESPACE`.

Closes iamacoffeepot/aether#654
